server/requirements.txt
absl-py==2.0.0
addict==2.4.0
astunparse==1.6.3
Authlib==1.3.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.6
decorator==4.4.2
dill==0.3.8
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.1
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lapx==0.5.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
onnxsim==0.4.35
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
proglog==0.1.10
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
setuptools==57.4.0
shapely==2.0.2
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf==1.0.0
thop==0.1.1.post2209072238
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.225
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.14.1
Pillow